requirements.txt
streamlit
clickhouse-connect
supabase
python-dotenv
requests
pandas
plotly
